refactor: cleanup module expand
