+++ src/expand_field_split.c
@@ -13,6 +13,47 @@
 #include "mod_expand.h"
 
 /**
+ * @brief Splits an expanded node into several words.
+ *
+ * Count the words with ft_count_expand_words().
+ * If one word return.
+ * If no word delete the node.
+ * If more words:
+ * Setup buffer with ft_init_buf().
+ * Split into several words with ft_split_node().
+ * Del old node and update list pointer with ft_del_old_node().
+ * @param list Pointer pointer to the current node.
+ * @param words Where to save the count.
+ * @return t_err SUCCESS, ERR_MALLOC.
+ */
+t_err	ft_field_split(t_tkn_list **list, size_t *words)
+{
+	t_tkn_list	*tmp;
+	t_buf		buf;
+	t_err		err;
+
+	tmp = *list;
+	ft_count_expand_words(tmp->content, words);
+	if (*words == 1)
+		return (SUCCESS);
+	else if (*words == 0)
+		ft_del_node_mid(&tmp);
+	else
+	{
+		err = ft_init_buf(&buf);
+		if (err != SUCCESS)
+			return (err);
+		err = ft_split_node(&tmp, &buf);
+		free(buf.str);
+		if (err != SUCCESS)
+			return (err);
+		ft_del_old_node(&tmp, words);
+	}
+	*list = tmp;
+	return (SUCCESS);
+}
+
+/**
  * @brief Iterate over a quoted part of string.
  *
  * @param quote_start Pointer where opening quote was found.
@@ -105,46 +146,5 @@ t_err	ft_del_old_node(t_tkn_list **list, size_t *words)
 	while (i++ < *words)
 		(*list) = (*list)->prev;
 	ft_del_node_mid(list);
-	return (SUCCESS);
-}
-
-/**
- * @brief Splits an expanded node into several words.
- *
- * Count the words with ft_count_expand_words().
- * If one word return.
- * If no word delete the node.
- * If more words:
- * Setup buffer with ft_init_buf().
- * Split into several words with ft_split_node().
- * Del old node and update list pointer with ft_del_old_node().
- * @param list Pointer pointer to the current node.
- * @param words Where to save the count.
- * @return t_err SUCCESS, ERR_MALLOC.
- */
-t_err	ft_field_split(t_tkn_list **list, size_t *words)
-{
-	t_tkn_list	*tmp;
-	t_buf		buf;
-	t_err		err;
-
-	tmp = *list;
-	ft_count_expand_words(tmp->content, words);
-	if (*words == 1)
-		return (SUCCESS);
-	else if (*words == 0)
-		ft_del_node_mid(&tmp);
-	else
-	{
-		err = ft_init_buf(&buf);
-		if (err != SUCCESS)
-			return (err);
-		err = ft_split_node(&tmp, &buf);
-		free(buf.str);
-		if (err != SUCCESS)
-			return (err);
-		ft_del_old_node(&tmp, words);
-	}
-	*list = tmp;
 	return (SUCCESS);
 }
